requirements.txt
yfinance
python-telegram-bot==20.7
